refactor: 클라이언트 및 서버 컴포넌트 분리
